+++ packages/terra-slide-panel/tests/wdio/slide-panel-spec.js
@@ -1,5 +1,18 @@
 /* global browser, Terra, before */
 
+describe('Large size squished slide panel', () => {
+  before(() => browser.setViewportSize(Terra.viewports('large')[0]));
+  beforeEach(() => {
+    browser.url('/#/tests/slide-panel-tests/behavior-squish-size-large');
+  });
+
+  Terra.should.matchScreenshot();
+  Terra.should.beAccessible();
+  Terra.should.themeEachCustomProperty({
+    '--terra-slide-panel-panel-width-large-scalar': '0.8',
+  });
+});
+
 describe('Toggle the slide panel and hidden styles', () => {
   before(() => browser.setViewportSize(Terra.viewports('large')[0]));
   beforeEach(() => {