+++ packages/terra-slide-panel/CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 Unreleased
 ----------
+### Fixed
+* Css bug in calcuated squish behavior
 
 1.16.0 - (February 1, 2018)
 ------------------